fix: consolidate and standardize workflows
