+++ .github/workflows/backport-merge-label.yml
@@ -19,16 +19,20 @@ jobs:
     timeout-minutes: 10
     steps:
       - name: 'Checkout Repository'
+        id: checkout-repository
         # https://github.com/actions/checkout/releases
         uses: actions/checkout@de0fac2e4500dabe0009e67214ff5f5447ce83dd # v6.0.2
         with:
           fetch-depth: 0
           ref: 'main'
       - name: 'Find and Label Related Issue'
+        id: find-label-issue
         # https://github.com/actions/github-script/releases
         uses: actions/github-script@3a2844b7e9c422d3c10d287c895573f7108da1b3 # v9.0.0
+        env:
+          SCRIPT_MODE: merge-label
         with:
           script: |
-            const scriptPath = `${process.env.GITHUB_WORKSPACE}/.github/workflows/scripts/merge-label.js`;
+            const scriptPath = `${process.env.GITHUB_WORKSPACE}/.github/workflows/scripts/backports.js`;
             const { default: script } = await import(scriptPath);
             await script({github, context, core});
