--- .github/workflows/scripts/clear-runner.js
@@ -1,53 +1,0 @@
-import { execSync } from 'child_process';
-export default async ({ core }) => {
-  const pathsToRemove = [
-    '/usr/lib/jvm',
-    '/usr/share/dotnet',
-    '/usr/share/swift',
-    '/usr/local/.ghcup',
-    '/usr/local/julia*',
-    '/usr/local/lib/android',
-    '/usr/local/share/chromium',
-    '/opt/microsoft',
-    '/opt/google',
-    '/opt/az',
-    '/usr/local/share/powershell',
-    '/opt/hostedtoolcache'
-  ];
-
-  try {
-    const output = execSync(`df -h / --total | grep total | awk '{print $4}'`).toString();
-    core.info(`Available disk space before cleanup: ${output}`);
-  } catch (error) {
-    throw new Error(`Failed running df to see disk space: ${error}`);
-  }
-
-  // Iterate over paths and remove them
-  for (const path of pathsToRemove) {
-    core.info(`Removing ${path}...`);
-    try {
-      // We use 'bash -c' to ensure wildcards (like julia*) are expanded correctly
-      execSync(`sudo bash -c "rm -rf ${path}"`);
-    } catch (error) {
-      throw new Error(`Failed to remove ${path}: ${error}`);
-    }
-  }
-
-  core.info('Pruning Docker...');
-  try {
-    execSync(`docker system prune -af`);
-    execSync(`docker builder prune -af`);
-    execSync(`docker image prune -af`);
-    execSync(`docker volume prune -af`);
-  } catch(error) {
-    throw new Error(`Failed pruning Docker: ${error}`);
-  }
-
-  core.info('Disk space after cleanup:');
-  try {
-    const output = execSync(`df -h / --total | grep total | awk '{print $4}'`).toString();
-    core.info(`Available disk space after cleanup: ${output}`);
-  } catch (error) {
-    throw new Error(`Failed running df to see disk space: ${error}`);
-  }
-};
