+++ package.json
@@ -1,7 +1,7 @@
 {
   "devDependencies": {
     "@eslint/js": "^9.39.3",
-    "eslint": "^9.39.3"
+    "eslint": "^10.8.0"
   },
   "type": "module"
 }
